project/requirements.txt
aiogram==3.x
python-dotenv
tortoise-orm
APScheduler
loguru
aiohttp
